requirements.txt

# Core Flask dependencies
flask>=3.1.2
flask-login>=0.6.3
flask-migrate>=4.1.0
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.2
werkzeug>=3.1.3

# Database
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.43
dj-database-url>=3.0.1

# Image processing
pillow>=11.3.0

# Payment gateways
stripe>=12.5.1
midtransclient>=1.4.2

# PDF and Excel generation
reportlab>=4.4.4
openpyxl>=3.1.5

# Environment and validation
python-dotenv>=1.1.1
email-validator>=2.3.0

# Django chat service dependencies
django>=5.2.6
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.2
redis>=5.0.1

# JWT for authentication
pyjwt>=2.8.0
cryptography>=41.0.8

# HTTP requests for service communication
requests>=2.31.0
email-validator
flask
flask-login
flask-migrate
flask-socketio
flask-sqlalchemy
flask-wtf
midtransclient
openpyxl
pillow
psycopg2-binary
pyjwt
python-dotenv
reportlab
requests
sqlalchemy
stripe
werkzeug
